Move LedgerDateStoreProvider into the shared folder
