+++ src/components/ChartOfAccounts/ChartOfAccounts.tsx
@@ -3,7 +3,7 @@ import { type ReactNode, useContext } from 'react'
 import { useChartOfAccounts } from '@hooks/legacy/useChartOfAccounts'
 import { useLedgerAccounts } from '@hooks/legacy/useLedgerAccounts'
 import { useElementViewSize } from '@hooks/utils/size/useElementViewSize'
-import { LedgerDateStoreProvider } from '@providers/LedgerDateStore/LedgerDateStoreProvider'
+import { LedgerDateStoreProvider } from '@providers/DateStoreProvider/LedgerDateStoreProvider'
 import { ChartOfAccountsContext } from '@contexts/ChartOfAccountsContext/ChartOfAccountsContext'
 import { InAppLinkProvider, type LinkingMetadata } from '@contexts/InAppLinkContext'
 import { LedgerAccountsContext } from '@contexts/LedgerAccountsContext/LedgerAccountsContext'
