+++ src/components/JournalTable/JournalTableWithPanel.tsx
@@ -3,8 +3,8 @@ import { CirclePlus } from 'lucide-react'
 import { useTranslation } from 'react-i18next'
 
 import { useSizeClass } from '@hooks/utils/size/useWindowSize'
+import { useLedgerDateRange } from '@providers/DateStoreProvider/LedgerDateStoreProvider'
 import { useJournalNavigation } from '@providers/JournalStore/JournalStoreProvider'
-import { useLedgerDateRange } from '@providers/LedgerDateStore/LedgerDateStoreProvider'
 import { JournalContext } from '@contexts/JournalContext/JournalContext'
 import { Button } from '@ui/Button/Button'
 import { Heading } from '@ui/Typography/Heading'
